backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
Pillow==11.1.0
